relocate search to ne wmodules as documented folder structure
