+++ app/frontend/src/modules/search/__mocks__/location.js
